daemon: allow loading app manifests from string
Currently, app sideloading requires an app manifest to be present
in the local filesystem. This means that each manifest has to be
written into a temporary file first, and then again copied to the
local manifest store. Besides possible security implications (i.e.
blindly writing any content to the local, temporary filesystem),
this produces avoidable overhead, especially when sideloading
from the WebApp.

This patch allows loading an app manifest directly from a string
instead, which gives the opportunity to validate an app manifest
YAML before writing it to disk.