daemon: prepare apps module as replacement for app_manager
Currently, app_manager handles everything related to apps and
instances. With v2 of the FLECS API, it should be split up into
two separate parts - one for apps and one for instances.

Both of them will implement asynchronous calls and progress-
tracking through the 'jobs' API, which is also t.b.d.

This 'apps' module implements API v2 for app management. Note that
until the 'instances' module is done as well the FLECS API will
not behave as expected. Same goes for interaction through the
WebApp.