daemon: implement GET methods of jobs API
Asynchronous operations each track their progress in a job_progress
structure. Through a centralized API a list of all jobs as well as
the progress of a single job should be retrievable.

This commit implements said /jobs and /jobs/{id} routes and adds
improvements to the underlying structures, such as defined start
values and automatically setting/resetting some status flags in
the progress structure.