daemon: prepare instances module to replace app_manager
Currently, app_manager handles everything related to apps and instances.
With v2 of the FLECS API, it should be split up into two separate parts -
one for apps and one for instances.

Both of them will implement asynchronous calls and progress- tracking
through the 'jobs' API, which is currently in early draft state.

This 'instances' module will implements API v2 for instance management.
Note that completion of this module is required for the FLECS API to
behave correctly. Same goes for interaction through the WebApp.