daemon: fix potential crash if instance cleanup is unsucessful
In case an instance remains throughout uninstallation of an App
(e.g. as instance deletion is unsuccessful or not performed due
to shutdown ...), creating a new instance after reinstallation of
this App should pick up the leftover instance and reactivate it.
However, this currently leads to a crash as pointers are not
checked in the according code path.

Instead of direct pointer access, use the safe interface provided
by the instance, which returns a valid result even if the instance
is currently orphaned.