build: run git version detection through sudo
Git 2.35.2 introduced new security measures, which prohibit
execution of git commands in a repository owned by another
user. It is therefore no longer possible to generate the
GIT_SHA appended to our software version. As a solution,
detect the owner of the root .git directory und use sudo
to run the GIT_SHA generation as this user.

As this does not work for our github actions, security
is explicitly disabled for them, as it does not add any
value in the context of throw-away containers.