deployment: fix backup of config files on up-/downgrade
Upon moving Instances to a newer/older version of the App, a backup is
created to ensure the user can go back to a previous version without
loss of data.

Two issues exist with this backup: First, Backing up of config files
is attempted to the wrong directory. Second, the backup directory is not
created, leading to permanently failing backups.

Fix backup path and create it in the process.