flecsport: nasty hack to upload imports to disk instead of ram
crow does not seem to have any nice way to directly upload files to
disk. For imports, however, this seems quite important as smaller
devices may quickly run out of ram when imports ranging in the GB
region are uploaded.

Hack something into their http parser as a - hopefully - temporary
workaround. In the future, this should be discussed with the author(s)
of crow and made into a nice feature, if possible.