flecsport: copy archive to volume-restore container
When running FLECS in a containerized context, mapping archives for
extraction is not always a valid option to restore volumes. This is
because extraction takes place in a new container context, which may
have a different view on directories (namely /var/lib/flecs/imports).
The safer approach is to copy the archives to the new container context
to make sure they exist in the expected location.