jobs: map instance_id to App in job descriptions
For async operations on Instances, only the instance_id is shown in the
job description (such as "stopping instance xyz"). To provide more
information to the user, also the App should be shown, in case the user
wants to configure, remove, ... this specific Instance.

Change the relevant operations to display the App alongside the
instance_id (such as "stopping instance xyz (App1 1.2.3.4-1)")